New command "Z" saves an instant of your index and working in the stash.
This is different from regular stashing because there is no prompting for a
name, and the stash is immediately re-applied.  It's just a way of taking a
quick snapshot of your working tree, so you can come back to later if need
be.

For example, if you had just made a lot of changes to your project, and then
needed to make another sweeping change before you could commit again, this
would let you safely store your changes without having to resort to a branch.