Fix magit-get-top-dir from ruining paths with symlinks.
One might argue that this also makes it work better in a case when GIT_DIR
points outside of a work tree, but because in that case any directory is a
valid "top dir" from GIT's point of view, that is debatable.  But at least
one can use Magit in that case by first making sure that default-directory
is the top dir and then executing magit-status.