Filter out HEAD refs from magit-name-rev.
HEAD refs are nasty because they're highly contingent.
Branches, tags, and remotes all require some commit to be made
in order for the ref to point to a different rev,
but HEADs just need something else to be checked out.

Also, filtering out HEADs makes the name more portable.
This is useful for Magithub, where the name may be passed on to GitHub.