A better way for updating the VC modeline
Revision 51a1b6d8 introduced a change that reverts all buffers after a
commit operation for refreshing the VC modeline info. This can be slow
(because refontification, etc) and have side effects (changes on display
settings for some modes.) Instead of reverting the buffer, we now call
`vc-find-file-hook' which is much faster and have no side effects. We
now update the VC modeline info after resetting, because that operation
may change the VC status without changing the file contents (commit a
change and then reset to HEAD^, for instance.)