in Makefile, fix the way we set emacs load-path
When using the native Windows emacs build (not a cygwin build), but a cygwin
build of 'make', then make fails because we send unix-style paths using CURDIR
to emacs.  If instead we use (expand-file-name "."), then emacs does the magic
for us.

I looked around and found that org-mode's Makefile does exactly this.  It seems
like the right thing to do.