Optionally use pop-to-buffer instead of switch-to-buffer
Define a new symbol magit-buffer-switch that's aliased to
pop-to-buffer by default. This is especially useful for people with
large wide monitors, who'd prefer all Magit actions happening on an
alternate buffer. For people with smaller screens, this symbol should
be aliased to switch-to-buffer. Leave the pop-to-buffer in
magit-log-edit as it is.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>