Do not always return t from magit-section-case
magit-guess-branch do use value returned by the clause in the
magit-section-case but magit-section-case just throw away those value,
and return t, making magit-guess-branch to failed.

So magit-section-case will return the value returned by the clause if
it is non-nil. But it must return t when the value is nil, otherwise
it will break the extension mechanism.