Support for a path spec in log mode with "=p"
Things to note: The path spec is not an option but actually two: "--"
followed by the path spec. Therefore the functions building the
arguments had to be adjusted accordingly.

Also ido-read-file-name throws errors if it is called from a key group
menu even if Emacs is started without any user or site configuration:

(wrong-type-argument symbolp (lambda nil (interactive) (magit-key-mode-add-argument (quote logging) "Path spec" "--" (quote magit-key-mode-read-file-name))))

Therefore magit-key-mode-read-file-name makes sure to unset
read-file-name-function if it is set to ido-read-file-name. This is a
hack until the root cause has been found and fixed.