Branch manager: Optionally remove branches in a remote repository or on request
Uses the usual: git push some-origin :refs/heads/some-branch

When removing a remote branch in a local repository magit will ask the
user whether or not she wants to remove the corresponding branch in
the remote repository as well.

Can be triggered by "K" in the branch manager as well. In this case
magit queries the user for the names of the remote and the branch in
that remote. It is therefore independent from point and from the
knowledge about the actual branches in the remote repository.

The remote itself must have been configured before.