Run git asynchronously from a rebase step
Rationale: An interrupted rebase (due to conflicts) will cause git to
open the editor so that the user can change the commit message when
the user continues the rebase ("R c" in magit). If the editor invoked
by Git is emacsclient then it is highly likely that the same Emacs
instance running magit is also used by emacsclient.

So far git has been running synchronously from
magit-rebase-step. Unfortunately this causes emacsclient to hang as
well.