Don't put the text at the end of the log buffer in a section.
There's no reason for that text to be in a section, and with the way the section was getting created it wound up being the first target of `magit-goto-next-section` when you were at the beginning of the log buffer.  Since it wasn't a commit section, that resulted in an error.