Automatically determine version within the GIT repository.
The version number is set to the output of `git describe --tags
--dirty`.  If the current HEAD is tagged (as released version are),
then that will be the name of the tag.  Otherwise it will be the tag
name, plus the number of commits on top of that tag, plus a short hash
of the exact commit, plus "-dirty" if there are changes in the working
tree.  Thus, the version number will actually tell us the exact
version being used when a development snapshot is installed

The Makefile's `dist` and `install` targets will substitute in the
proper version numbers when they create files to be used outside the
Git directory.

bin/mk_rel.bash has been simplified greatly, since there's no longer
any consistency checking for it to do.