Use `magit-section`s in the branch manager buffer
Also make `magit-show-branches-mode` derive from `magit-mode`.

Besides consistency, the principle benefits are that the
`magit-item-highlight` face appears on the branch that commands will
act on, and that key combinations (e.g. `l l`) which work in other
Magit buffers will work in the branch manager buffer as well.

I rewrote the `magit-show-branches` command so that it inserts text in
sections instead of adding text properties, and changed a few utility
functions to get information from section info instead of the old text
properties.

I also replaced the `magit-branches-window-checkout` command with a
new action in `magit-visit-item` for visiting a branch.