Make Typechange diffs behave reasonably in the status buffer.
Expanding a Typechange diff with `TAB` reveals two "sub-diffs".  E.g.,
a deleted file and a new symlink of the same name.  (The commit buffer
and the `git diff` command always show them as two files like that.)
When attempting to stage/unstage/discard a sub-diff, act as though we
were staging/unstaging/discarding the parent diff.