Fix bug deleting wrong remote branch
When attempting to delete a remote-tracking branch and the associated
remote branch with `k` in the branch manager, the code prior to this
change would:
1) Delete the branch at point.
2) Refresh the buffer as a side effect of calling Git.
3) Delete the remote branch referred to by the branch at point.

Of course, the branch at point in #3 is NOT the same branch that was
at point when the user pressed `k`.  I introduced this bug when I
taught the branch manager's buffer how to respond to magit-refresh.

To fix it, I capture the branch information in a variable before
making any changes, and use that variable instead of any *-at-point
functions.  I also don't change anything until after asking the user
if he wants to remove the remote branch as well, so the selected
branch is still visible when the question is asked.