Don't touch line-move-visual and truncate-lines
`truncate-lines` was set to make the diffs display better.  That's not
necessary with colored diffs.

`line-move-visual` was set "primarily to work around some bugs
somewhere related to hiding and showing of text."

Whatever bugs were being worked around don't seem to be an issue now,
so leave these two variables as user settings.