Rebase mode: Show keybindings for editing commands
The comment that Git puts at the bottom of the file shows all the
commands with their one-letter abbreviation.  All the one-letter
abbreviations correspond to rebase-mode's default keybinding for that
command, except for `p`, which is bound to `previous-line`.  It's not
obvious what key to press for `pick`.

Now, change the comment to show the key bound to each command in place
of the abbreviation.  By default, that just means changing a 'p' to a
'c', but if a user has redefined the keybindings for whatever reason,
the actual binding will be shown for all commands.