Make refresh keep the point on the correct line even when the point is not at the beginning of the line.
If you hit refresh (g) while the point is not at the beginning of the line,
the point will end up on the line below the one it is supposed to be on.

This is due to the way count-lines work. From the documentation:
  (count-lines START END)

  Return number of lines between START and END.
  This is usually the number of newlines between them,
  but can be one more if START is not equal to END
  and the greater of them is not at the start of a line.

If the point is not at the beginning of the line, the following expression is off by one:
(count-lines (magit-section-beginning old-section)
	     (point))