don't apply a region of a hunk with no context
When experimenting with applying an active region of a hunk, I found
that the results could be quite confusing when the hunk had zero
context.  This patch modifies magit-apply-hunk-item* to signal an
error in that situation.

 - Added local variables to reduce some code duplication in deciding
   if the context lines are zero, and if we are using the region or
   not.

 - Added documentation

I didn't want to change it with this commit, but it appears to me that
the REVERSE argument to magit-apply-hunk-item* is equivalent to
checking for "--reverse" in ARGS.  REVERSE as is, I think, is a little
hard to understand since it's only used in if the region is
active *because* it's only needed explicitly to pass to the code that
creates the diff for the region.