Got rid of magit--branch-view-details.
This function cluttered magit-branch-manager.
It forced to pass around an associative list and to extract info with (too
many) assoc-default.

Got rid of most of the information that was added to a branch section.
It seems that only 'branch and 'remote are ever used.
Got rid of magit--branch-view-details.

This function cluttered magit-branch-manager.
It forced to pass around an associative list and to extract info with (too
many) assoc-default.

Got rid of most of the information that was added to a branch section.
It seems that only 'branch and 'remote are ever used.