Got rid of 'remote in the branch info.
This was a flag just saying if the branch at point was remote.
This was not needed, as the 'branch part of the branch info is enough to reconstruct it (by checking for "^remotes").

This was not standard. All the other info have only one piece of information.
This none-standardization meant that assoc-default was used to retrieve the information.
Commit to be merged before publishing.
Reverting "Adding a fetch-at-point function for the branch manager." Adding a fetch-at-point function for the branch manager.