Remove ":require magit-wip" from define-minor-mode
It can cause problems if you load your custom file before loading
magit, which is quite likely. I think the :require option is really
only for core Emacs features.

As a workaround, I call the setup function when the file is loaded,
which will do the setup if the mode variable is true.