Makes it easy to add wildcards to git ignore files.
This commit proposes to enrich the history in the minibuffer with
wildcard expressions of the current filename. For example, for a
filename `/myfolder/file.o', the following options are proposed in the
minibuffer history (accessible with M-n):

/myfolder/file.o
/myfolder/*.o
*.o
file.o

The options order depends on the negativity of the prefix argument.

This commit additionally provides comments for related functions.

Signed-off-by: Damien Cassou <damien.cassou@gmail.com>