Set before-save-hook to nil in rebase-mode
Rebase mode becomes unusable if a function in `before-save-hook`
throws an error.  This could happen, for example, with
`delete-trailing-whitespace`, since the buffer is read-only.

By doing this in a hook function instead of directly, we make it easy
for the user to stop it from happening if they know what they're doing
and they have a `before-save-hook` that really makes sense in the
context of rebase-mode.

This is intended as a cleaner way of solving the same bug as
d4a3dff143efbecd62f86b2bf2b307a74a6473cd.

In consultation with: Ryan C. Thompson <rct@thompsonclan.org>