Preference for status comments in log edit buffer
I've added a preference that, if set to t, puts the standard message
"Please enter the commit message..." that normally appears when you do
"git commit" on the command line, along with the output of "git
status". This works well when magit-log-edit-mode is derived from
"git-commit-mode", which does nice font-locking on the inserted
comments. Hence I have put it in the git-commit-mode branch.

The hook function and variable have the same name right now, and
that's a problem but I'm not creative enough to come up with distinct
names at 3:00 AM. Please suggest better names.

See below for an example of what gets inserted by this hook:

 # Please enter the commit message for your changes. Lines starting
 # with '#' will be ignored, and an empty message aborts the commit.
 # On branch git-commit-mode
 # Changes to be committed:
 #
 #	modified:   magit.el
 #