Fix default rev in magit-rebase-step
Before, it would default to a remote tracking branch in the format
"origin (master)", even if `magit-remote-ref-format` was set to
`remote-slash-branch`.  This means that if the user accepted the
default value, the command would fail because of the badly formatted
ref.

Now, use the same logic to put default value in the "user-friendly
format" that `magit-read-rev` will later use to translate it into the
canonical form.