Honor git commit.template config setting.
Add a hook to magit-log-edit-mode to insert the contents of the
commit.template file if configured.  This appears to match the
behavior of git from the command line.
As far as I can tell, git does the following:
If the commit.template is configured in a .git/config, treat a relative
path as relative to the top directory of the repo.
If it is a global config (~/.gitconfig) or an absolute path, use that path.

Using this, every time I open magit-edit-log, I get this I'm not sure
about that, but it's hidden in *Messages*:

    Buffer `*magit-edit-log*' not visiting a file