Request to take down the gh-pages
Hi Phil,

You have in your github fork of github a gh-pages branch that is served by github. It is coming up really high in search engines results. The problem is that it is out of date and does not redirect easily to the current manual.

Could you remove that branch from your github repository to avoid the confusion for new users?

Thanks in advance,
Nicolas