Add option to auto-advance to the next line after selecting action
It's convenient in some situations to auto-advance point to the next
line after selecting an action for a given commit in the rebase-todo
buffer (after all, presumably you're done selecting the action for
that line). Add an option to automatically advance after selecting an
action.

This cuts out one keystroke for the common scenario of selecting an
action moving to the next line.