Uses a colon separated refspec for pull
`git pull origin master` does not store the information retrieved from
the remote in the local copy of the remote (refs/remotes/origin/master).
It uses FETCH_HEAD instead to store the information temporarily and then
does the merge. As a consequence, when comparing refs/heads/master and
refs/remotes/origin/master, it looks like the local branch is ahead and
it magit will display that there are some unpushed commits.

To alleviate the problem it is necessary to use the following command:
`git pull origin refs/heads/master:refs/remotes/origin/master`.