git grep should use --untracked by default
Rationale:




Short version: If we wanted git-grep to ignore any files, they would already be in .gitignore




Long version: Generally the only time we have untracked files is either if we forgot to add them to .gitignore (but will soon), or if they're legitimate files that we want to consider part of the project but just haven't committed yet. It's confusing and unexpected behavior for git-grep (at least within magit) to not notice occurrances of a search within files that you just created, until shortly after you remember that they just aren't tracked by git yet.