Introduce magit-log-simple for graph-less log
magit-log is only useful when navigating recent history.  To navigate
larger histories, it is imperative to turn off the graph and increase
the default number of entries displayed.  Introduce a magit-log-simple
that does exactly this.  Also teach magit-key-mode.el about it, and
switch contrib/magit-simple-keys.el to use it by default.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>