commit-mode.el: initial import of git-commit-mode
git-commit-mode [1] is a great mode for editing git commit messages,
but is unfortunately maintained poorly.  Import it into the magit tree
as commit-mode.el (similar to rebase-mode.el), so we can have tighter
magit integration and better maintenance.  Subsequent commits will
modify the file significantly and integrate it into
magit-log-edit-mode.

[1]: https://github.com/rafl/git-commit-mode

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>