commit-mode.el: remove git-define-git-commit macro
As opposed to git-define-commit-mode-self, which just uses the current
user's name and email, this macro prompts for a name and email on the
minibuffer.  This is unnecessary, as the user is perfectly capable of
editing text after it is filled into the buffer.  Also migrate its
callers to use git-define-commit-mode-self.

While at it, give git-define-commit-self a more sensible name:
commit-mode-define-header.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>