magit-wip-save: do not attempt to save gitignored files
Previously an attempt was made but git-wip failed to save the change.

If there is a ignore rule that matches the file being saved but it is
actually being tracked or even just in the (real, not the tempory one
used for the wip branch) index then it *is* being saved successfully,
as before.