Fix bug where the wrong author info is used for committing
The author information is saved in the buffer-local variable 'process-environment' in
the log-edit buffer.
To reset its value, kill-local-variable must be called not after, but BEFORE
switching away from the log-edit buffer.

Otherwise, the old author information will be reapplied in the next commit that uses
the same log-edit buffer.