magit-cmd-output: do not corrupt output
This fixes #544.  There is no reason why we would want to manipulate
git's output.  In the master and next branches something similar to
the code removed in this commit was used to have colorful graphs in
logs.  But doing it in `magit-cmd-output' was wrong and that has been
fixed in these branches also.