tests/run-tests.el: add /path/to/magit/../{mocker,cl-lib} to load-path
It is quite reasonable to expect the directories containing the
dependencies to be located in the same directory as the magit
directory is located in.  At least this is the case when manually
installing each package as a subdirectory of some "site-lisp"
directory.  Or when using submodules as I do.