Simplify finding pseudo-header position
The logic git-commit-find-pseudo-header-position is unnecessary buggy
and unnecessarily convoluted.  It currently makes the assumption that
only pseudo-header lines can contain the ':' character.  This faulty
assumption manifests itself as bugs when attempting to "C-c C-s" in a
commit message like this one (where the summary contains a ':').

Further, this is unnecessarily convoluted as there is no need to look
for existing pseudo headers at all.  It's sufficient to look for the
first non-empty line preceding the comments (if any).

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>