Fix buggy insert-header logic
The current logic for inserting the header after determining header-at
is riddled with bugs.  Most significantly, it completely omits the case
when header-at is at end-of-line (which happens when your buffer is not
terminated with a newline).  Also, the logic for looking for an existing
pseudo-header is completely bogus: it just looks for a line with the
character ':'.

Redo the entire logic in a dedicated function.  This function uses
git-commit-known-pseudo-headers to find existing pseudo headers
correctly and can return either zero, one, or two newlines.  Its logic
is flawless, as it considers every corner case.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>