Makefile: don't mention flet is obsolete when compiling
There were many warnings when byte-compiling magit, that have been
ignored in the past.  I suspect that is because the output was to noisy
and the warnings were just overlocked.  This change reduces the noise
level by not mentioning that flet is obsolete (we know).  Combined with
the other noise reduction I have done in recent commits, I hope this
will allow us to catch and fetch compiler warnings sooner.