fix #667 without adding additional cruft
The problem with my original patch wasn't (just) that I forgot to
account for the possibility of the second argument unintentionally
containing format parameter but that using it as a format string was
wrong altogether.

`magit-insert-status-line' now takes exactly two arguments the HEADING
and the INFO-STRING and inserts both literally without processing them
using `format' first.

There was _one_ instance where we could previously write
  (magit-insert-status-line "heading" "%s %s" (a) (b))
and now have to write
  (magit-insert-status-line "heading" (concat (a) " " (b)))

If there were more instances like this then using the second argument
format string _might_ be justified (more likely not: it still makes it
harder to understand `magit-insert-status-line' without actually
thinking).  As it stands using a format string just obfuscates the
functions usage, which goes against the initial motivation of making the
code easier to read pleasing to the eye.

Furthermore if we need to document this function's use despite it being
very simple then there is something wrong with it.  Now a doc-string is
not necessary anymore, because I removed the initial "design" error and
use more descriptive argument lines.