when running tests: ensure previous git process is done
When the new variable `magit-process-force-synchronous' is non-nil
`magit-run*' runs all processes synchronously.  This is necessary
because when running tests on travis it can happen that the previous
git process hasn't finished yet when an attempt is made to creating
the next one.  This causes tests to fail randomly on travis, depending
on general load.

This new variable should only be used for its intended purpose.