fix magit-item-highlight for default theme
In the default theme both `highlight' (which `magit-item-highlight' used
to inherit from) and the various `magit-diff-*' faces set the background
color.  Making the former override the latter, which then have no effect
whatsoever.

Since our faces have to make sense when using the default theme we have
no choice but to use some other face attribute for `magit-item-highlight',
even though there is a good change that this will break some third-party
themes.

So in case people start complaining, please tell them to contact the
maintainer of their themes instead and point them to this commit.

This finally fixes issue #133.