magit-key-mode-command: don't quote arguments to git
Git is invoked directly, there is no shell involved that would need the
the quoting to prevent splitting an argument containing whitespace into
multiple arguments.

In most cases the quoting is unnecessary when the argument contains
shell globs (to be interpreted by git not a shell) quoting did change
the globs, which is not what we want.