Add variable for default log range
Using git log without specifying a range is very slow on some git
repositories (like the Linux kernel). It's tedious to have to specify
a custom range every time you need to look at the log. Add a variable
for the default log range that can be set to things like
"HEAD~100..HEAD", or to a function that returns a similar string.

Left unchanged, this variable introduces no user-visible changes.

Example usage of specifying a function that returns a range string:
https://github.com/mgalgs/.emacs.d/blob/f9ae63829d87a7b3842bc7d15563b633169ec613/magit-setup.el